Add: CRUD clients

+++ cerveceria/src/main/java/com/cocobongo/cerveceria/sales/dto/RegisterSaleRequest.java
@@ -23,7 +23,7 @@ public class RegisterSaleRequest {
     private PaymentType paymentType;
  
     // Nullable — solo requerido si paymentType = CREDIT
-    private Integer clientId;
+    private Long clientId;
  
     @NotEmpty(message = "La venta debe incluir al menos un producto")
     @Valid                   // propaga validación a cada SaleItemRequest
